+++ examples/authentication_riverpod/lib/main.dart
@@ -24,7 +24,7 @@ class MyApp extends HookConsumerWidget {
             return container.read(authProvider).status ==
                 AuthStatus.authenticated;
           },
-          beamToNamed: '/login'),
+          beamToNamed: (_, __) => '/login'),
 
       /// if the user is anything other than authenticated
       /// else send them to /home
@@ -35,7 +35,7 @@ class MyApp extends HookConsumerWidget {
             return container.read(authProvider).status !=
                 AuthStatus.authenticated;
           },
-          beamToNamed: '/home'),
+          beamToNamed: (_, __) =>'/home'),
     ],
     initialPath: '/login',
     locationBuilder: (routeInformation, _) => BeamerLocations(routeInformation),